Removed DataFile from inheritance structure, all comments for FileDataset expectations moved to file_dataset.py
