+++ great_expectations/dataset/__init__.py
@@ -1,4 +1,4 @@
-from .base import Dataset, DataFile
+from .base import Dataset
 from .datatable import Datatable
 from .pandas_datatable import MetaPandasDatatable, PandasDatatable
 from .sqlalchemy_datatable import MetaSqlAlchemyDatatable, SqlAlchemyDatatable
