+++ great_expectations/dataset/base.py
@@ -1052,338 +1052,6 @@ If you wish to change this behavior, please set discard_failed_expectations, dis
 
 
 
-class DataFile(Dataset):
-    @classmethod
-    def file_lines_map_expectation(cls, func):
-
-        """Constructs an expectation using file lines map semantics.
-
-        The file_lines_map_expectations decorator handles boilerplate issues surrounding the common pattern of \
-        evaluating truthiness of some condition on an line by line basis in a file.
-
-        Args:
-            func (function): \
-                The function implementing an expectation that will be applied line by line across a file \
-                The function should take a file and return information about how many lines met expectations.
-
-        Notes:
-            Users can specify skip value k that will cause the expectation function to disregard the first
-            k lines of the file
-
-        See also:
-            :func:`expect_file_line_regex_match_count_to_be_between <great_expectations.dataset.base.Dataset.expect_file_line_regex_match_count_to_be_between>` \
-            for an example of a file_lines_map_expectation
-        """
-        raise NotImplementedError
-
-
-
-
-    def expect_file_line_regex_match_count_to_equal(self,regex, lines=None,expected_count=0, skip=None,
-                                                mostly=None, result_format=None,
-                                                include_config=False,catch_exceptions=None,meta=None):
-        """
-        Expect the number of times a regular expression appears on each line of a file to equal some
-        expected count
-
-        Args: regex:
-            A string that can be compiled as valid regular expression
-
-        Keyword Args:
-            lines: \
-                An empty variable that recieves the file lines from the file_lines_map_expectation method.
-                It doesn't matter what the user gives for this value as it will be replaced with the lines
-                of the file by file_lines_map_expecation. It is recommended the user ignore this argument and leave
-                it at its default.
-
-            skip (nonnegative integer): \
-                Integer specifying the first lines in the file the method should skip before assessing
-                expectations
-
-            expected_count (None or nonnegative integer): \
-                Specifies the number of times regex is expected to appear on each line of the
-                file
-
-
-            mostly (None or number between 0 and 1): \
-
-                Specifies an acceptable error for expectations. If the percentage of unexpected lines is
-                less than mostly, the method still returns true even if all lines don't match the
-                expectation criteria.
-
-            result_format (str or None): \
-                Which output mode to use: `BOOLEAN_ONLY`, `BASIC`, `COMPLETE`, or `SUMMARY`.
-                For more detail, see :ref:`result_format <result_format>`.
-            include_config (boolean): \
-                If True, then include the expectation config as part of the result object. \
-                For more detail, see :ref:`include_config`.
-            catch_exceptions (boolean or None): \
-                If True, then catch exceptions and include them as part of the result object. \
-                For more detail, see :ref:`catch_exceptions`.
-            meta (dict or None): \
-                A JSON-serializable dictionary (nesting allowed) that will be included in the output without modification. \
-                For more detail, see :ref:`meta`.
-
-        Returns:
-
-            A JSON-serializable expectation result object.
-
-            Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
-            :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
-
-        """
-        raise NotImplementedError
-
-    def expect_file_line_regex_match_count_to_be_between(self,regex,lines=None, skip=None,
-                                                         expected_min_count=0, expected_max_count=None,
-                                                         mostly=None, result_format=None, include_config=False,
-                                                         catch_exceptions=None, meta=None):
-
-        """
-        Expect the number of times a regular expression appears on each line of a file to be between a
-        maximum and minimum value.
-
-        Args: regex:
-            A string that can be compiled as valid regular expression
-
-        Keyword Args:
-            lines: \
-                An empty variable that recieves the file lines from the file_lines_map_expectation method.
-                It doesn't matter what the user gives for this value as it will be replaced with the lines
-                of the file by file_lines_map_expecation. It is recommended the user ignore this argument and leave
-                it at its default.
-
-            skip (nonnegative integer): \
-                Integer specifying the first lines in the file the method should skip before assessing
-                expectations
-
-            expected_min_count (None or nonnegative integer): \
-                Specifies the minimum number of times regex is expected to appear on each line of the
-                file
-
-            expected_max_count (None or nonnegative integer): \
-               Specifies the maximum number of times regex is expected to appear on each line of the
-               file
-
-            mostly (None or number between 0 and 1): \
-
-                Specifies an acceptable error for expectations. If the percentage of unexpected lines is
-                less than mostly, the method still returns true even if all lines don't match the
-                expectation criteria.
-
-            result_format (str or None): \
-                Which output mode to use: `BOOLEAN_ONLY`, `BASIC`, `COMPLETE`, or `SUMMARY`.
-                For more detail, see :ref:`result_format <result_format>`.
-            include_config (boolean): \
-                If True, then include the expectation config as part of the result object. \
-                For more detail, see :ref:`include_config`.
-            catch_exceptions (boolean or None): \
-                If True, then catch exceptions and include them as part of the result object. \
-                For more detail, see :ref:`catch_exceptions`.
-            meta (dict or None): \
-                A JSON-serializable dictionary (nesting allowed) that will be included in the output without modification. \
-                For more detail, see :ref:`meta`.
-
-        Returns:
-
-            A JSON-serializable expectation result object.
-
-            Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
-            :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
-
-        """
-        raise NotImplementedError
-
-    def expect_file_hash_to_equal(self, value, hash_alg='md5', result_format=None,
-                                                    include_config=False,catch_exceptions=None,meta=None):
-        """
-        Expect computed file hash to equal some given value.
-
-        Args:
-            value: A string to compare with the computed hash value
-
-        Keyword Args:
-            hash_alg (string):  Indicates the hash algorithm to use
-
-            result_format (str or None): \
-                Which output mode to use: `BOOLEAN_ONLY`, `BASIC`, `COMPLETE`, or `SUMMARY`.
-                For more detail, see :ref:`result_format <result_format>`.
-
-            include_config (boolean): \
-                If True, then include the expectation config as part of the result object. \
-                For more detail, see :ref:`include_config`.
-
-            catch_exceptions (boolean or None): \
-                If True, then catch exceptions and include them as part of the result object. \
-                For more detail, see :ref:`catch_exceptions`.
-
-            meta (dict or None): \
-                A JSON-serializable dictionary (nesting allowed) that will be included in the output without modification. \
-                For more detail, see :ref:`meta`.
-
-        Returns:
-            A JSON-serializable expectation result object.
-
-        Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
-        :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
-
-        """
-        raise NotImplementedError
-
-    def expect_file_size_to_be_between(self, minsize, maxsize,result_format=None,
-                                                    include_config=False,catch_exceptions=None,
-                                                    meta=None):
-        """
-        Expect file size to be between a user specified maxsize and minsize.
-
-        Args:
-            minsize(integer): minimum expected file size
-            maxsize(integer): maximum expected file size
-
-        Keyword Args:
-
-            result_format (str or None): \
-                Which output mode to use: `BOOLEAN_ONLY`, `BASIC`, `COMPLETE`, or `SUMMARY`.
-                For more detail, see :ref:`result_format <result_format>`.
-
-            include_config (boolean): \
-                If True, then include the expectation config as part of the result object. \
-                For more detail, see :ref:`include_config`.
-
-            catch_exceptions (boolean or None): \
-                If True, then catch exceptions and include them as part of the result object. \
-                For more detail, see :ref:`catch_exceptions`.
-
-            meta (dict or None): \
-                A JSON-serializable dictionary (nesting allowed) that will be included in the output without modification. \
-                For more detail, see :ref:`meta`.
-
-        Returns:
-            A JSON-serializable expectation result object.
-
-        Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
-        :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
-        """
-
-        raise NotImplementedError
-
-    def expect_file_to_exist(self,result_format=None,include_config=False,
-                             catch_exceptions=None, meta=None):
-
-        """
-        Checks to see if a file specified by the user actually exists
-
-
-        Keyword Args:
-
-            result_format (str or None): \
-                Which output mode to use: `BOOLEAN_ONLY`, `BASIC`, `COMPLETE`, or `SUMMARY`.
-                For more detail, see :ref:`result_format <result_format>`.
-
-            include_config (boolean): \
-                If True, then include the expectation config as part of the result object. \
-                For more detail, see :ref:`include_config`.
-
-            catch_exceptions (boolean or None): \
-                If True, then catch exceptions and include them as part of the result object. \
-                For more detail, see :ref:`catch_exceptions`.
-
-            meta (dict or None): \
-                A JSON-serializable dictionary (nesting allowed) that will be included in the output without modification. \
-                For more detail, see :ref:`meta`.
-
-        Returns:
-            A JSON-serializable expectation result object.
-
-        Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
-        :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
-        """
-
-        raise NotImplementedError
-
-    def expect_file_to_have_valid_table_header(self,regex,skip=None,
-                                           result_format=None,
-                                           include_config=False,
-                                           catch_exceptions=None,meta=None):
-
-        """
-        Checks to see if table-like data file has unique column names
-
-
-        Keyword Args:
-            skip (nonnegative integer): \
-                Integer specifying the first lines in the file the method should skip before assessing
-                expectations
-
-            regex (string):
-                A string that can be compiled as valid regular expression. Used to specify the elements
-                of the table header (the column headers)
-
-            result_format (str or None):
-                Which output mode to use: `BOOLEAN_ONLY`, `BASIC`, `COMPLETE`, or `SUMMARY`.
-                For more detail, see :ref:`result_format <result_format>`.
-
-            include_config (boolean): \
-                If True, then include the expectation config as part of the result object. \
-                For more detail, see :ref:`include_config`.
-
-            catch_exceptions (boolean or None): \
-                If True, then catch exceptions and include them as part of the result object. \
-                For more detail, see :ref:`catch_exceptions`.
-
-            meta (dict or None): \
-                A JSON-serializable dictionary (nesting allowed) that will be included in the output without modification. \
-                For more detail, see :ref:`meta`.
-
-        Returns:
-            A JSON-serializable expectation result object.
-
-        Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
-        :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
-        """
-        raise NotImplementedError
-
-
-    def expect_file_to_be_valid_json(self, schema=None,result_format=None,
-                               include_config=False,catch_exceptions=None,
-                               meta=None ):
-
-        """
-        schema : string
-            optional JSON schema file on which JSON data file is validated against
-
-        result_format (str or None):
-            Which output mode to use: `BOOLEAN_ONLY`, `BASIC`, `COMPLETE`, or `SUMMARY`.
-            For more detail, see :ref:`result_format <result_format>`.
-
-        include_config (boolean):
-            If True, then include the expectation config as part of the result object. \
-            For more detail, see :ref:`include_config`.
-
-        catch_exceptions (boolean or None):
-            If True, then catch exceptions and include them as part of the result object. \
-            For more detail, see :ref:`catch_exceptions`.
-
-        meta (dict or None):
-            A JSON-serializable dictionary (nesting allowed) that will
-            be included in the output without modification. \
-
-        For more detail, see :ref:`meta`.
-
-        Returns:
-            A JSON-serializable expectation result object.
-
-        Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
-        :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
-        """
-        raise NotImplementedError
-
-
-
-
-
-
-
 
 ValidationStatistics = namedtuple("ValidationStatistics", [
     "evaluated_expectations",
